+++ src/protocol/mcpe/packet/s_to_c_handshake/handshake.rs
