+++ src/protocol/mcpe/packet/s_to_c_handshake/mod.rs
@@ -1,4 +1,4 @@
-pub mod key_exchange;
+pub mod handshake;
 pub mod constants;
 
 use anyhow::Result;
